Enforce docs in aztecjs and accounts packages
