+++ yarn-project/foundation/.eslintrc.cjs
@@ -1,41 +1,3 @@
-// See https://typescript-eslint.io/play/#ts=5.1.6&showAST=es&fileType=.ts
-const contexts = [
-  // All methods in an interface
-  'TSInterfaceDeclaration TSMethodSignature',
-  // All public methods in a class that does not implement an interface
-  'ClassDeclaration[implements.length=0] MethodDefinition[accessibility=public]',
-  // TODO: All methods public by default in a class that does not implement an interface
-  // 'ClassDeclaration[implements.length=0] MethodDefinition[accessibility=undefined][key.type=Identifier]',
-  // TODO: All export const from the top level of a file
-  // 'ExportNamedDeclaration[declaration.type=VariableDeclaration]',
-  // Legacy contexts (needs review)
-  'TSParameterProperty[accessibility=public]',
-  'TSPropertySignature',
-  'PropertySignature',
-  'TSInterfaceDeclaration',
-  'InterfaceDeclaration',
-  'TSPropertyDefinition[accessibility=public]',
-  'PropertyDefinition[accessibility=public]',
-  'TSTypeAliasDeclaration',
-  'TypeAliasDeclaration',
-  'TSTypeDeclaration',
-  'TypeDeclaration',
-  'TSEnumDeclaration',
-  'EnumDeclaration',
-  'TSClassDeclaration',
-  'ClassDeclaration',
-  'TSClassExpression',
-  'ClassExpression',
-  'TSFunctionExpression',
-  'FunctionExpression',
-  'TSInterfaceExpression',
-  'InterfaceExpression',
-  'TSEnumExpression',
-  'EnumExpression',
-];
-
-const JSDOC_RULES_LEVEL = 'off';
-
 module.exports = {
   extends: [
     'eslint:recommended',
@@ -112,27 +74,8 @@ module.exports = {
     ],
     'import/no-extraneous-dependencies': 'error',
     'import/no-cycle': 'warn',
-    'tsdoc/syntax': JSDOC_RULES_LEVEL,
-    'jsdoc/require-jsdoc': [
-      JSDOC_RULES_LEVEL,
-      {
-        contexts,
-        checkConstructors: false,
-        checkGetters: true,
-        checkSetters: true,
-      },
-    ],
-    'jsdoc/require-description': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-hyphen-before-param-description': [JSDOC_RULES_LEVEL],
-    'jsdoc/require-param': [JSDOC_RULES_LEVEL, { contexts, checkDestructured: false }],
-    'jsdoc/require-param-description': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-param-name': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-property': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-property-description': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-property-name': [JSDOC_RULES_LEVEL, { contexts }],
-    'jsdoc/require-returns': 'off',
     // this unfortunately doesn't block `fit` and `fdescribe`
     'no-only-tests/no-only-tests': ['error'],
   },
-  ignorePatterns: ['node_modules', 'dest*', 'dist', '*.js', '.eslintrc.cjs'],
+  ignorePatterns: ['node_modules', 'dest*', 'dist', '*.js', '.eslintrc.cjs', '.eslintrc.*.cjs'],
 };
